project name change, sql migration down for snippet model change fixed, logging middleware removed, home html fixed, uuid at postgres level for post id instead of serial numbers, mock model fixes, data model refactor according to change

+++ internal/store/mocks.go
@@ -20,7 +20,7 @@ func (m *MockSnippetStore) Insert(ctx context.Context, s *Snippet) (int, error) 
 	return 2, nil
 }
 
-func (m *MockSnippetStore) Get(ctx context.Context, id int) (*Snippet, error) {
+func (m *MockSnippetStore) Get(ctx context.Context, id int64) (*Snippet, error) {
 	switch id {
 	case 1:
 		return &m.Snippet, nil
